customer_support_api/requirements.txt
Flask>=3.0.0,<4.0.0
Flask-SQLAlchemy>=3.1.0,<4.0.0
Flask-JWT-Extended>=4.6.0,<5.0.0
Flask-Cors>=4.0.0,<5.0.0
Flask-Limiter>=3.5.0,<4.0.0
marshmallow>=3.21.0,<4.0.0
bcrypt>=4.1.0,<5.0.0
bleach>=6.1.0,<7.0.0
email-validator>=2.1.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
Werkzeug>=3.0.0,<4.0.0
